Fix Bugs

+++ Application/Features/TenantStudents/Dtos/StudentStatisticsDto.cs
@@ -3,7 +3,7 @@
     public sealed class StudentStatisticsDto
     {
         public int Students { get; set; }
-        public double AverageGrade { get; set; }
+        public int AverageGrade { get; set; }
         public int AttendanceRate { get; set; }
         public int ActiveStudents { get; set; }
     }
